+++ pkg/util/validation/blocked_request.go
@@ -3,7 +3,12 @@
 package validation
 
 type BlockedRequest struct {
-	Path        string            `yaml:"path,omitempty"`
-	Method      string            `yaml:"method,omitempty"`
-	QueryParams map[string]string `yaml:"query_params,omitempty"`
+	Path        string                              `yaml:"path,omitempty"`
+	Method      string                              `yaml:"method,omitempty"`
+	QueryParams map[string]BlockedRequestQueryParam `yaml:"query_params,omitempty"`
+}
+
+type BlockedRequestQueryParam struct {
+	Value    string `yaml:"value"`
+	IsRegexp bool   `yaml:"is_regexp,omitempty"`
 }